Refactor Login redirection
